Merged all changes
This is a pretty big update. I managed to bring together all of our parts and tested them. I also added audio and changed from our original idea of using a sword to hurling fireballs instead. This project is now a working prototype. Our menu system still needs some attention and we need to implement ways for the player to get hurt so it's actually possible to lose.